Fall back to the `guest.ipaddress` field if `guest.net` doesn't have ips
Govmomi `WaitForIp` only waits for the guest.ipaddress field to be
populated. We're using `guest.net` to get ips from the network cards.
This change adds a fallback to the `ipaddress` field if the network
cards don't have ips. This is likely when the VM has just booted up
because the `WaitForIp` method in the VMware library only waits for the
`ipaddress` field to be populated.